requirements.txt
accelerate==0.23.0
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.5
dbus-python==1.2.18
deepspeed==0.10.3
distro==1.7.0
filelock==3.12.4
fsspec==2023.9.2
hjson==3.1.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
lit==17.0.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
peft==0.4.0
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==1.10.12
python-dotenv==1.0.0
PyGObject==3.42.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
sympy==1.12
evaluate
datasets
matplotlib
wandb
black
scikit-learn
isort
sentencepiece
pynvml
pandas
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.3
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.5